backend/requirements.txt
# Flask core
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
Flask-Migrate>=4.0.5
Flask-CORS>=4.0.0
Flask-JWT-Extended>=4.6.0

# Database
SQLAlchemy>=2.0.36
alembic>=1.14.0

# Utilities
python-dotenv==1.0.0
